Refactor unet inference

+++ common/visualization.py
@@ -2,15 +2,13 @@ from typing import List, Tuple
 from PIL import Image
 import matplotlib.pyplot as plt
 import matplotlib.patches as patches
-from matplotlib.patches import Polygon as MplPolygon
-from sympy import Polygon
 
 
 def compare_images(
     images: List[Image.Image],
     titles: List[str],
     images_per_column: int = 3,
-    rect_coords: Tuple[float, float, float, float] | None = None,
+    rect_coords: Tuple[int, int, int, int] | None = None,
     zoom: bool = True
 ) -> None:
     assert len(images) >= len(titles), "More titles than images!"
@@ -18,7 +16,7 @@ def compare_images(
 
     image_x, image_y = images[0].size
     if rect_coords is None:
-        rect_coords = (image_x * 0.25, image_y * 0.25, image_x * 0.05, image_y * 0.05)
+        rect_coords = (int(image_x * 0.25), int(image_y * 0.25), int(image_x * 0.05), int(image_y * 0.05))
 
     # Calculate the number of rows needed for the given images per column
     rows = len(images) // images_per_column + (len(images) % images_per_column > 0)
@@ -44,7 +42,7 @@ def compare_images(
                 rect_coords[0], rect_coords[1], 
                 rect_coords[0] + rect_coords[2], 
                 rect_coords[1] + rect_coords[3]
-            )) # type: ignore
+            ))
 
             ax_inset = ax.inset_axes([0.7, 0.0, 0.3, 0.3])
             ax_inset.imshow(cropped_image, cmap="gray")
